+++ js/main.js
@@ -2,8 +2,8 @@ import { renderPictures } from './rendering-pictures.js';
 import {
   onUploadInputChange,
   onUploadCancelButtonClick,
-  validateUploadForm,
 } from './img-upload-form.js';
+import { validateUploadForm } from './validate-upload-form.js';
 
 onUploadInputChange();
 onUploadCancelButtonClick();
